+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## main
 
+- Add `erldns_questions` questions filter to the packet pipeline.
 - Update dns_erlang v4.2 and remove `erldns_records:name_type/1`.
 
 ## v7.0.0